elf: read-in objects contents at once
This implies we no longer need to heap allocate new strtab in Object.
Instead, we are storing pointers to strtab and shstrtab slices in
Object.contents. Furthermore, there is no need to store file_offset
for navigating the Object when part of Archive - since we read
everything in one go, we just need it for the initial read.