macho: filter symbols by section id first, then by address
This fixes a tricky bug when the compiler emits an empty section
with a symbol in the symtab. Without sorting by section id first,
we would mess up linking by assuming the lone symbol being in a different
section, or worse taking it in place of another symbol at the same
address belonging to the right section.