fully support indirect function table
Previously the table wouldn't be emitted properly into the final
binary. Also when the function table was to be imported, it would
try to find a definition from one of the object files. We now
correctly generate a new import for it with the correct amount of
function pointers. When no function pointers exist, we do not import
or generate a function table.