wasm: type merging - add index to dirty map
When we add a type to the list of function types, we set the
corresponding index to the function. We must add the function
index to the dirty map to ensure we do not try to merge the same
function type twice, as the original function type index has already
been updated and therefore points to the index within the new list.