wasm: fix resolution when both symbols are undef.
When both symbols are undefined, we keep the existing symbol but also
append the newly found symbol to the list of discarded symbols. This
would previously prevent the archive logic from finding newly found
undefined symbols which were defined in the archive file itself.

This also fixes writing the data section which was missing 0's in case
the last atom was unresolved and would break the loop.