update to latest master build of Zig
Updates to the new build API as well as the new for-loop syntax.
Syntax changes have been performed by running 'zig fmt'.
The build changes no longer support depending on the same file within
more than 1 module. To prevent having to fully restructure the test
suite, we set the main_pkg_path to the root directory as to allow for
imports outside the 'src' directory.