wasm: ensure `__wasm_init_memory` will be called
When we have shared-memory, the memory will be initialized during
module initialization. To ensure this is being done, we set the
mentioned function as the entrypoint of the WebAssembly module.
This function will be called upon thread creation too, to ensure
the memory is correctly initialized. The execution of the first
startup will then continue with `__start` (when available).