elf: redo DT_TEXTREL test to have text relocs in a DSO and not EXE
Previously, following mold, this test tested text relocs present in a PIE executable.
However, as we want to cover musl AND glibc, it is now modified to test presence of
text relocs in a DSO which is then linked with an executable.
According to Rich and this thread https://www.openwall.com/lists/musl/2020/09/25/4
musl supports only a very limited number of text relocations and only in DSOs (and
rightly so!).