wasm: handle data -and debug symbol liveness
For atoms which are not represented by a symbol, we create a synthetic
symbol instead. This way we are sure the symbol will get resolved and
is picked up in the garbage collection process. This also ensures we
do not attempt to generate symbols from atoms which had previously
no symbol, causing UB in unsafe build modes.