wasm: mark debug atoms with alive references
When a debug atom contains relocations to symbols which are considered
alive, we mark the debug atom's symbol alive also. This ensures the atom
will be allocated and therefore emit. When a relocation occurs within
a debug atom to a function that's considered dead, we reset the PC rather
than attempt to find its offset into the file.