wasm: Use Atom indexes rather than pointers
This means we do not have to allocate each Atom independently and can
directly store the Atom in the `managed_atoms` list. When we require
a pointer, we can simply get a pointer into the list. This is, however,
less safe as a new allocation could move the list. This change does
allow us to use less memory and slightly improved perf due to the
removal of the allocations.