wasm: write linked binary to buffer before file
Rather than performing each write operation into the file, we
first write it to a temporary buffer. The syscalls to allocate memory
has found to be cheaper than the syscall to perform a file write.
From testing a fully linked binary went from 0.30s to 0.26.

In the future we could look into precalculating the size of certain
sections and allocating enough for such section to reduce the amount
of allocations done.