wasm: refactor object parsing into atoms
Instead of parsing all sections into atoms, we are now given a symbol
for which we will parse its data into an atom. This allows us to only
parse it for symbols which are referenced and are alive, saving us
from unneccarily parsing sections which are not referenced by any
symbols.