wasm: parsing atom fixes
We now assign the relocations directly to the `Atom`. Previously we
would edit the original relocation's offset, which would make it end up
as part of other atoms also. We now instead store the original offset
of an atom within the section of the object file and substract the
value upon relocation rewrite. This does use 4 bytes extra per Atom
with regards to memory, but does mean we do not have to duplicate
relocations, which would cost even more memory.

This also fixes retrieving the correct code section for function symbols
as previously it assumed incorrect indexes due to imported functions.

Lastly, we do no longer try to parse atoms for undefined symbols, which
obviously do not have an associated atom.