wasm: sort code section pre-allocation
The code -and function sections must match in order within the binary.
Previously we would order the code section before writing them to disk.
However, this meant they were already allocated and the offsets were
based on the previous order. This meant that debug info was incorrect.
We now order the atoms before allocation, and ensure synthetic functions
are created after allocation, but appended correctly to the code section
to ensure they are emit last, and therefore have correct offsets.