wasm: fix function offset relocations
When calculating the function offset for its relocation we would
previously use atom's offset with a fixed additional offset. However, we
must include the size of the previous function's body which is LEB128-
encoded. This means we cannot use a fixed-size offset to calculate the
function body offset within the code section unless we use a fixed-size
LEB size, which would increase the binary size. Instead, we simply re-
calculate the atom's offset based on the currently written bytes during
atom writing as we will not need this offset until we perform relocations
for the debug section anyway.