wasm: order and allocate code atoms on write
Rather than ordering the atoms of the code section earlier, we simply
skip it until we write the actual code section. This is possible because
we don't need the know the offset of each atom until we perform the
relocations of the debug sections, which we already delay to writing of
those sections. Debug sections *must* always come after the module,
including the code section. Therefore this is fine to relay on.
We now do a lot less work and make the codebase simpler as well.
This also allows us to remove the `next` field on atoms, reducing the
memory usage every so slightly.